Add files via upload
This project is used to send the mail to the desired person (the admin) by the user who is basically sending the mail through the contact form. This project uses Embedded JavaScript (ejs) as the view engine. All you have to do is just type in your correct gmail username and password in the app.js page.